server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.8.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filetype==1.2.0
Flask==3.0.3
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
google-genai==1.45.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0
langchain-core==1.0.0
langchain-google-genai==3.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
sentry-sdk==2.42.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.0.4
xxhash==3.6.0
zstandard==0.25.0
